requirements.txt
# IMU Interfacing
adafruit-circuitpython-bno08x-rvc
adafruit-circuitpython-bno08x
Adafruit-Blinka
adafruit-circuitpython-busdevice
RPi.GPIO
pyserial
pose3d

# Camera Interfacing
opencv-python

# Recording data
mcap
protobuf
mcap-protobuf-support